Add ESLint with Airbnb defaults
Set up ESLint, and use the Airbnb base configuration.

The Etherpad hook API has a few quirks that are not compatible with
Airbnb's style guide. The main problem is that the names of the hooks
are a mix of camelCase and snake_case, such as "eejsBlock_exportColumn".
They do provide the ability to map hook names to our own internal names
via the configuration in ep.json[1], but I think our code will be easier
to understand if we use the proper hook names directly.

    You can omit the FUNCTIONNAME part, if the function to register has
    got the same name as the hook. So "authorize" :
    "ep_PLGUINNAME/YOURFILE" will call the function exports.authorize in
    ep_PLUGINRNAME/YOURFILE

So, we must override a few rules.

First, use named exports rather than default exports
(no-default-export[2], prefer-default-export[3]); this is both a
personal preference and is part of the Etherpad plugin API, as it will
not load default exports.

Second, disable the camelcase[4] rule: I think our code will be easier
to understand if we use hook names directly, despite being named with
occasional underscore. We could selectively override this for each hook
name, but I think that would be more work than is worthwhile.

Finally, the Etherpad hook API seems to require assigning to function
arguments, so we must disable the no-param-reassign rule. Again, we
could selectively override this rule for each such assignment, but
will not for now.

Also clean up the code style violations I already introduced in the six
lines of actual code in this repo.

[1] https://github.com/ether/etherpad-lite/wiki/Creating-a-plugin#epjson
[2] https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-default-export.md
[3] https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/prefer-default-export.md
[4] https://eslint.org/docs/rules/camelcase
[5] https://eslint.org/docs/rules/no-param-reassign