Save plugin settings in a promise
Getting settings from the Etherpad Lite server's settings.json file is a
surprisingly difficult task.

The main way I've seen, used by most of the most popular plugins, is to
reach into the server internals and import a utility file:

    const settings = require('../../src/node/utils/Settings');

or
    const settings = require('ep_etherpad-lite/node/utils/Settings');

I am worried that this approach would lead to plugins breaking if the
Etherpad Lite developers ever decided to change how they load settings.
It also has two additional drawbacks: it appears not to be compatible
with TypeScript's import statements, and there is a bug related to
developer installs[1].

The official method - used infrequently - is the loadSettings hook[2].
This has the drawback of being called after plugins are loaded[3], which
means that plugins need to defer some of their initialization until
after the hook is fired.

To handle that asynchronous gap and represent the "coming soon" status
of the settings at initialization, use a Promise[4]. This adds some
complexity to the use of these settings, as everything needs to be
wrapped in either an await or a .then(), but it seems to be the most
correct way to work with the Etherpad Lite server settings.

At this time, I am not validating that the settings block for the plugin
contains the expected structure; that is something that I want to do,
but I do not foresee any challenges in doing so, and so will turn my
attention to higher-risk items in the meantime.

[1] https://github.com/ether/etherpad-lite/issues/4047#issuecomment-634962074
[2] https://etherpad.org/doc/latest/#index_loadsettings
[3] https://github.com/ether/etherpad-lite/blob/f6df9ffad08e2dc17ad0391be88d0c7034b4e1c0/src/node/server.js#L137-L145
[4] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Using_promises