Validate debouncing pad updates works
The padUpdate hook[1] is called on every change to a pad, as its name
suggests. We don't want to sync to Permanent on literally every update;
instead, we want to wait until some amount of time has passed since the
last update.

Lodash provides the mechanism to do that: debounce[2]. Debounce will
group multiple updates, wait for the configured time to elapse, and then
invoke the callback with the most recent update.

This functionality is one of the things I expected to work, but wanted
to verify early, so for now, it's just logging to the console. The next
step here is to replace that with a call to the node-sdk to send the
contents to Permanent.

[1] https://etherpad.org/doc/latest/#index_padupdate
[2] https://lodash.com/docs/4.17.15#debounce