Add API endpoints
The front end will need to be able to get and set the Permanent export
configuration on each pad. Register the initial API endpoints for doing
so, including wiring up the ability to send JSON data as the request
body.

Since we're directly working with the Express API, dev-depend on the
type definitions and peer-depend on the package itself.

Example cURL requests:

    curl -v http://localhost:9001/p/demo/permanent

    curl -v --request PUT \
      -H "Content-Type: application/json" \
      --data '{"example": true}' \
      http://localhost:9001/p/demo/permanent