Prototype database storage
Set up a very simple initial example of getting data from the database
and storing storing data in the database.

The key bit of research here was the module declaration; it took me a
moment to realize that I didn't need to tell TypeScript about the base
module ('ep_etherpad-lite'), and instead needed to tell it about the
full import path ('ep_etherpad-lite/node/db/DB').

We will probably want to add better type definitions for the Etherpad
API later, but this is enough for us to move forward.