Wire up jQuery for use in prototype frontend form
Set up an initial prototype of the form users will use to turn on
syncing to their Permanent archive. In order for the form to correctly
send data to the Etherpad backend, and not do a full page load, we need
to attach an event handler to the form. Etherpad provides jQuery, which
we can use in the document ready hook to look up our element and attach
our handler.