Get settings via internal API instead of hook
The loadSettings hook[1] is, as best I can tell, insufficient for
receiving settings that need to be used in eejs hooks. The eejs hooks do
not allow async content, and the server appears to cache the hook so
they cannot be altered after loading (eg, cannot be altered in response
to loadSettings).

This internal API approach is the one used by pretty much every other
plugin that reads settings. It is a bit of a risk, as internal APIs can
change, but these are the tools available to us.

[1] https://etherpad.org/doc/latest/#index_loadsettings