List files to include in package
To prepare for publishing (in at least some form), we need to make sure
that `npm pack`[1] does the right thing. By default, it will exclude
everything in .gitignore[2], but that is the opposite of what we want:
we do want the built JavaScript, and don't really want the source
TypeScript files.

We could use .npmignore, but better to use the allow-list approach[3]
that the "files" field offers[4]:

    The optional `files` field is an array of file patterns that
    describes the entries to be included when your package is installed
    as a dependency.

Include our committed static assets, our uncommited built JavaScript,
the Etherpad plugin definition, and the README.

[1] https://docs.npmjs.com/cli/v6/commands/npm-pack
[2] https://docs.npmjs.com/cli/v6/using-npm/developers#keeping-files-out-of-your-package
[3] https://stackoverflow.com/a/37421736
[4] https://docs.npmjs.com/cli/v6/configuring-npm/package-json#files