Switch UI based on web monetization status
We want to gate the sync-to-Permanent functionality behind web
monetization being enabled. First, check to see if the COIL plugin is
installed; then, wait for the plugin to report that payments are
flowing.

There is a bit of a race condition we want to avoid here: it could be
the case that our document ready hook gets called before or after
payments begin flowing. We need to check the current state at the time
of hook execution to determine which case we're in, and if payments have
not yet started, register the callback for when they are.

The types-wm package declares TypeScript types for web monetization.
Since the web monetization API is through a property on the global
`document` object, we don't (yet) actually need any of the types it
declares, but we do need to import the package in order to get access to
that extended global. We can import an empty list of types in order to
satisfy both the compiler and the linter.

[1] https://www.npmjs.com/package/types-wm