Parse cookies on plugin API routes
As a first pass at auth, we'll be taking advantage of the fact that the
session cookies the Permanent application sets are available to
subdomains. This will allow us to sidestep most of the authentication &
authorization issues for the moment, until the Permanent application has
a better auth story.

Additionally, we can get the Etherpad author by way of the `token`
cookie[1], so we will need that as well.

Add the cookie-parser[2] middleware so that cookies are available to all
the request handlers in the plugin's API.

[1] https://etherpad.org/doc/latest/#index_cookies
[2] https://www.npmjs.com/package/cookie-parser