Show current sync status on front-end
Set up the data structure to be used in the database, and the API
response for getting the current sync status, and use that response to
show the correct bit of front-end UI: not logged in, not syncing, sync
pending, or syncing.

We're not yet doing anything to change the sync status, as that requires
further changes.