Build UI and API to enable and disable sync
Set up the API endpoints for a given Etherpad author to enable or
disable sync to their Permanent archive. The API is actually pretty
simple: POST to enable, and DELETE to disable.

Most of the complexity relates to properly handling invalid credentials,
where the user has the cookies that, perhaps, once came from Permanent,
but they are no longer valid. We can't know the credentials are invalid
until we try to use them (which we're not doing yet), and for the sake
of APIs that return in a reasonable time we won't be doing that during
the API call that enables sync. Once we do learn that the credentials
are invalid, we still need to record them in the database so that the
API can distinguish between "you still have the same invalid session,
and cannot enable sync" and "you have a new session, and can try again".

Update the data structures we'll be storing in the database to allow for
that distinction, implement the enable & disable endpoints, and call
them from the front-end.

This is a follow-up to commit 2b3034360a256718f6ceafb2a726d10cb13f0041,
which introduced a lot of these pieces.