Add Express as a dev dependency
Etherpad provides Express in the context the plugin executes, and we say
that the plugin uses Express by having it as a peer dependency. However,
in order for our tools to properly analyze import statements and types,
we also need it as a direct dependency, as the server context is
unavailable at build / lint time. Add Express as a dev dependency.