Add dev dependency on Etherpad instance
In order for ESLint to correctly resolve import references, we need to
have the Etherpad source directory linked. We already had it as a peer
dependency.

Unfortunately, as the Etherpad project does not publish this separately,
we must have a local reference to it. This dependency assumes that
Etherpad is checked out in a peer directory of the plugin repo:

    ./ep_permanent_exporter
    ./etherpad-lite