Validate credentials & upload to Permanent
Integrate with the node-sdk[1]. Validate the credentials by getting (or
creating) the 'Etherpad' folder in the Apps directory of the user's
default archive, and save the resulting information to the database to
use for subsequent initializations of the node-sdk.

Update the pad update hook debounce function to actually send the pad
text to Permanent, using the previously validated credentials & target
directory.

To be able to upload according to each author's sync settings, we need
to fetch all the sync configs for a given pad. UeberDB2 offers a way to
find all the matching keys for a wildcard query[2], but not to also get
their values in a single query; unfortunately, that forces us into using
the n+1 query antipattern. I don't expect this to ever run at the scale
where that will be a signficant performance impact; if it is, we can
look at instead storing all authors' sync configs in a shared key.

[1] https://github.com/PermanentOrg/node-sdk/
[2] https://github.com/ether/ueberDB#findkeys