Remove debugging CSS border
It was helpful to have a big red border around the content the plugin
added to the export column to make sure that the element was loading
correctly. We no longer need it! The element is loading correctly, and I
don't anticipate any major changes to the layout in the near future, so
remove the red border.