Rewrite plugin to authenticate via OAuth
Replace the proof-of-concept cookie authentication, which required
deployment to a subdomain of permanent.org, with proper authentication
as an OAuth client.

Instead of saving cookies in the database to be used for later uploads
triggered by changing synced pads, save OAuth tokens. Change the data
model for sync configs to store pad => author and author => token,
rather than pad => cookie, so that an author can re-log-in after the
token has expired to re-enable syncing.