Work around Etherpad bug
Etherpad is doing some cleverness around allowing require statements
across the network. Unfortunately, if the file the code running in the
browser is trying to require ends in a comment, part of the
network-loader-modified code is commented out, and the whole thing
becomes unparsable.

Building a source map adds a comment to the end of the compiled code
with a reference to that source map, triggering the bug.

Turn off source maps in our TypeScript configuration.